Allow defining priority label key and reading priority from labels (#2392)
### Description of change


![screencapture-localhost-3000-d-1cafa24eadc87e0eeb9217b9179e31473bcfa21d-aperture-service-protection-2023-07-18-09_01_29](https://github.com/fluxninja/aperture/assets/1553055/c5880956-ab83-4abc-b995-f154571b2034)

##### Checklist

- [x] Tested in playground or other setup
- [x] Screenshot (Grafana) from playground added to PR for 15+ minute
run
- [x] Documentation is changed or added
- [x] Tests and/or benchmarks are included
- [ ] Breaking changes

+++ pkg/scheduler/request.go
@@ -2,13 +2,13 @@ package scheduler
 
 // Request is metadata for request in a flow that is to be allowed or dropped based on controlled delay and queue limits.
 type Request struct {
-	FairnessLabel string // for enforcing fairness
-	Tokens        uint64 // tokens (e.g. expected latency or complexity) for this request
-	InvPriority   uint64 // larger values represent higher priority
+	FairnessLabel string  // for enforcing fairness
+	Tokens        uint64  // tokens (e.g. expected latency or complexity) for this request
+	InvPriority   float64 // larger values represent higher priority
 }
 
 // NewRequest calculates the inverse priority and returns a new Request.
-func NewRequest(fairnessLabel string, tokens uint64, invPriority uint64) *Request {
+func NewRequest(fairnessLabel string, tokens uint64, invPriority float64) *Request {
 	return &Request{
 		FairnessLabel: fairnessLabel,
 		Tokens:        tokens,
